Solved "AttributeError: 'list' object has no attribute 'to' "
There was an error when you ran the batch command. This is due to the updated transformers library. The issue is resolved by adding a return_tensors point. Just to be safe, I've left out the old batch command as well so those who are running the old library can also use it (it is commented out though)